fix: :lock: 提高MISRA-C:2012标准的符合度

+++ xwmd/libc/newlibac/mem.c
@@ -77,7 +77,7 @@ void * _calloc_r(struct _reent * r, size_t elem_nr, size_t elem_sz)
         rc = xwmm_mempool_malloc(newlibac_mempool, total, &mem);
         errno = -rc;
         if (XWOK == rc) {
-                memset(mem, 0, total);
+                memset(mem, 0, total); // cppcheck-suppress [misra-c2012-17.7]
         }
         return mem;
 }
@@ -99,9 +99,9 @@ int posix_memalign(void ** memptr, size_t alignment, size_t size)
 {
         int rc;
 
-        if (((alignment & (alignment - 1)) != 0) ||
-            (alignment % sizeof(void *) != 0) ||
-            (alignment == 0)) {
+        if (((alignment & (alignment - (size_t)1)) != (size_t)0) ||
+            (alignment % sizeof(void *) != (size_t)0) ||
+            (alignment == (size_t)0)) {
                 rc = EINVAL;
         } else {
                 void * mem;
@@ -172,6 +172,7 @@ void _malloc_stats_r(struct _reent * r)
         XWOS_UNUSED(r);
 
         mi = mallinfo();
+        // cppcheck-suppress [misra-c2012-17.7]
         fprintf(stderr, "max system bytes = %10lu\n", (long) mi.arena);
         /* FIXME */
 }
